refactor: Modal 내부에서 useFieldArray를 직접 선언하고 관리하도록 구현
